refactor: move buttons to separate directory in widget folder
